Move thread out of core
