+++ ak-py/src/agentkernel/thread/model.py
